Move worktree tests to their own file
